+++ comics/accounts/urls.py
@@ -4,7 +4,8 @@ from django.views.generic.simple import direct_to_template
 
 from registration.views import activate, register
 
-from comics.accounts.forms import AuthenticationForm
+from comics.accounts.forms import (AuthenticationForm, PasswordChangeForm,
+    PasswordResetForm)
 
 urlpatterns = patterns('',
     url(r'^register/$',
@@ -48,7 +49,10 @@ urlpatterns = patterns('',
 
     url(r'^password/change/$',
         auth_views.password_change,
-        {'template_name': 'auth/password_change.html'},
+        {
+            'template_name': 'auth/password_change.html',
+            'password_change_form': PasswordChangeForm,
+        },
         name='auth_password_change'),
     url(r'^password/change/done/$',
         auth_views.password_change_done,
@@ -62,6 +66,7 @@ urlpatterns = patterns('',
             'email_template_name': 'auth/password_reset_email.txt',
             # XXX The next line is not supported until Django 1.4
             #'subject_template_name': 'auth/password_reset_email_subject.txt',
+            'password_reset_form': PasswordResetForm,
         },
         name='auth_password_reset'),
     url(r'^password/reset/confirm/(?P<uidb36>[0-9A-Za-z]+)-(?P<token>.+)/$',