+++ comics/comics/romanticallyapocalyptic.py
@@ -15,7 +15,6 @@ class Crawler(CrawlerBase):
     def crawl(self, pub_date):
         page = self.parse_page('http://www.romanticallyapocalyptic.com/')
         urls = page.src('img[src*="/art/"]', allow_multiple=True)
-        for possible_url in urls:
-            if 'thumb' not in possible_url:
-                url = possible_url
-        return CrawlerImage(url)
+        for url in urls:
+            if 'thumb' not in url:
+                return CrawlerImage(url)