+++ comics/core/urls.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from django.conf.urls.defaults import *
 
 from comics.core import views
@@ -47,8 +46,3 @@ urlpatterns = patterns('',
     # We do not like robots
     url(r'^robots.txt$', views.robots, name='robots'),
 )
-
-if settings.DEBUG:
-    urlpatterns += patterns('',
-        url(r'^widgets/$', views.widgets, name='widgets'),
-    )