+++ docs/changes.rst
@@ -36,6 +36,7 @@ v1.1.3 (in development)
 
 **Crawlers**
 
+- New: ``dresdencodak``
 - New: ``extraordinary``
 - New: ``icanbarelydraw``. Contributed by Jim Frode Hoff.
 - New: ``manlyguys``. Contributed by Jim Frode Hoff.
@@ -43,6 +44,7 @@ v1.1.3 (in development)
 - Update ``eatthattoast`` after site change.
 - Update ``extralife`` to be more robust.
 - Update ``questionablecontent`` to be more robust.
+- Update ``romanticallyapocalyptic`` to be more robust.
 - Update ``threewordphrase`` to be more robust.